La til regler så alle brukere kan logge ut, kun ansatte kan registrere sak, og it kan alle andre funksjoner

+++ main.py
@@ -11,11 +11,12 @@ from models.user import User
 def menu():
     system = SupportSystem()
     case_id = 1
-
+    # The user can choose to log in as either an employee or an IT admin
     print("Choose role:")
     print("1. Employee")
     print("2. IT Admin")
 
+    # Asking the user to select a role based on the menu options
     while True:
         role_choice = input("Select role: ")
         if role_choice == "1":
@@ -26,35 +27,47 @@ def menu():
             break
         else:
             print("Invalid choice, try again.")
-            
+
     # Loop keeps running until the user chooses to exit the program
     while True:
         print("\n--- IT SUPPORT SYSTEM ---")
         print(f"Logged in as: {current_user.name} ({current_user.role})")
-        print("1. Register case")
-        print("2. Show cases")
-        print("3. Update status")
-        print("4. Close case")
-        print("5. Delete case")
-        print("6. Close program")
+        print("0. Log out")
+
+        if current_user.is_employee():
+            print("1. Register case")
+
+        elif current_user.is_admin():
+            print("2. Show cases")
+            print("3. Update status")
+            print("4. Close case")
+            print("5. Delete case")
+        
         # User can select an option from the menu
         choice = input("Choose alternative: ")
 
-        # User can add a description of the issue and set the priority level
-        if choice == "1":
+        # Loggin out of the program for all users:
+        if choice == "0":
+            print("Log out...")
+            break
+
+        # For employees:
+        # Adds a description of the issue and set the priority level
+        elif choice == "1" and current_user.is_employee():
             description = input("Describe the issue: ")
             priority = input("Priority (Low/Medium/High): ")
             system.create_case(description, case_id, priority)
             print("Case registered!")
             case_id += 1
-
-        # Shows all existing tickets in the system
-        elif choice == "2":
+        
+        #For admins:
+        # Shows all existing tickets in the system to the IT admin
+        elif choice == "2" and current_user.is_admin():
             system.show_cases()
 
         # The user can update the status of a selected ticket
         # The user needs to enter a valid ID and a valid status to update 
-        elif choice == "3":
+        elif choice == "3" and current_user.is_admin():
             try:
                 case_id_input = int(input("Write ID: "))
                 case = system.find_case(case_id_input)
@@ -74,7 +87,7 @@ def menu():
                 print("Invalid input.")
 
         # Closes selected ticket by changing the status to "Closed"
-        elif choice == "4":
+        elif choice == "4" and current_user.is_admin():
             try:
                 case_id_input = int(input("ID: "))
                 case = system.find_case(case_id_input)
@@ -89,7 +102,7 @@ def menu():
                 print("Invalid input.")
 
         # Deletes a selected case
-        elif choice == "5":
+        elif choice == "5" and current_user.is_admin():
             try:
                 case_id_input = int(input("ID: "))
                 if system.delete_case(case_id_input):
@@ -98,11 +111,6 @@ def menu():
                     print("Did not find case.")
             except ValueError:
                 print("Invalid input.")
-
-        # Exits the program
-        elif choice == "6":
-            print("Program closing...")
-            break
 
         else:
             print("Invalid choice.")
